+++ src/main/java/org/rgomez/springboot/calculadora/services/OperationService.java
@@ -1,5 +1,7 @@
 package org.rgomez.springboot.calculadora.services;
 
+import java.math.BigDecimal;
+
 import org.rgomez.springboot.calculadora.request.Operador;
 import org.rgomez.springboot.calculadora.response.ResultResponse;
 
@@ -13,6 +15,6 @@ public interface OperationService {
 	 * @param number2
 	 * @return ResultResponse
 	 */
-	ResultResponse calcular(Operador operacion, double number1, double number2);
+	ResultResponse calcular(Operador operacion, BigDecimal number1, BigDecimal number2);
 
 }
